moved help_command.py from Models to view
