+++ src/test/test_imports.py
@@ -87,7 +87,6 @@ def test_import_entity():
     assert Entity.rename_field
     assert Entity.get_method
     assert Entity.add_method
-    assert Entity.add_method_and_params
     assert Entity.edit_method
     assert Entity.delete_method
     assert Entity.rename_method
@@ -104,7 +103,7 @@ def test_import_uml_method():
     assert UML_Method.change_parameters
 
 def test_import_help():
-    from src.umleditor.mvc_model.help_command import help_menu
+    from umleditor.mvc_view.help_command import help_menu
     assert help_menu
 
 def test_import_relation():
