+++ src/umleditor/mvc_model/__init__.py
@@ -4,5 +4,4 @@ sys.path.append('src/umleditor/mvc_model')
 from .custom_exceptions import *
 from .diagram import *
 from .entity import *
-from .help_command import *
 from .relation import *
