+++ src/umleditor/mvc_view/help_command.py
