--- src/umleditor/usability/__init__.py
@@ -1,5 +1,0 @@
-import sys
-sys.path.append('src/umleditor/usability')
-
-from .autofill import *
-from .momento import *
